move object svg
