+++ src/shapes/Object/FabricObject.ts
@@ -1,5 +1,5 @@
 import { ObjectEvents } from '../../EventTypeDefs';
-import { FabricObjectSVGExportMixin } from '../../mixins/object.svg_export';
+import { FabricObjectSVGExportMixin } from './FabricObjectSVGExportMixin';
 import { InteractiveFabricObject } from './InteractiveObject';
 import { applyMixins } from '../../util/applyMixins';
 import { StatefulMixin } from '../../mixins/stateful.mixin';
